Fix for requestHandler error case
This is the problem that was causing the bug:

In case there is an error while fetching data (just intentionally use a wrong URL to reproduce the issue locally, for example instead of "/tasks" use "/taskssss"), the error is caught in the fetchData method of requestHandler but App.js -which is calling fetchData- tries to apply a then block after catch. Since the catch block does not provide any value for the chain to continue, then block cannot be executed.

The solution I implemented is, I just moved to catch block from requestHandler.js to App.js to make it executed after all "then" blocks